Publish/Subscribe, Pre-compiled.
This version of the esp8266 client program the full program structure.SSL is enabled. The program performs publish and subscribe
to full topics with full messages that are extracted from the XML file.

The umqtt package (simple.py) is pre-compiled
and then inserted into the module.

The program runs without problems.