Included Severity Mode feature
Server checks for the severity mode of every
new client node in order to restore the node
to it's last known severity mode.

Also, on a severity mode change on the node the
server accepts the incoming change message from
the node and savesit's last severity mode on the database.